state-transition general
