+++ state-transition/general.json
@@ -254,7 +254,7 @@
   },
   {
     "id": 3,
-    "description": "2 accounts and 4 invalid transactions. Two of the transactions have invalid from and to address so no rawTx is provided for them, can be ignored.Tx with same amount than balance and tx with invalid chain id (different chain id than sequencer). Old root equals new root",
+    "description": "2 accounts and 2 invalid transactions. Tx with same amount than balance and tx with invalid chain id (different chain id than sequencer). Old root equals new root",
     "arity": 4,
     "chainIdSequencer": 1001,
     "defaultChainId": 1000,
@@ -288,26 +288,6 @@
         "rawTx": "0xf86e80843b9aca00830186a0944d5cf5032b2a844602278b01199ed191a86c93ff88016345785d8a000080820346a06e209c61ca92c2b980d6197e7ac9ccc3f547bf13be6455dfe682aa5dda9655efa016819a7edcc3fefec81ca97c7a6f3d10ec774440e409adbba693ce8b698d41f1",
         "customRawTx": "0xee80843b9aca00830186a0944d5cf5032b2a844602278b01199ed191a86c93ff88016345785d8a00008082019180806e209c61ca92c2b980d6197e7ac9ccc3f547bf13be6455dfe682aa5dda9655ef16819a7edcc3fefec81ca97c7a6f3d10ec774440e409adbba693ce8b698d41f11c",
         "reason": "TX INVALID: Chain ID does not match"
-      },
-      {
-        "id": 1,
-        "from": "0x617b3a3528F9cDd6630fd3301B9c8911F7Bf063D",
-        "to": "0x4d5Cf5032B2a844602278b01199ED191A86c93f",
-        "nonce": 0,
-        "value": "100000000000000000",
-        "gasLimit": 100000,
-        "gasPrice": "1000000000",
-        "chainId": 1001
-      },
-      {
-        "id": 2,
-        "from": "0x617b3a3528F9cDd6630fd3301B9c8911F7Bf03D",
-        "to": "0x4d5Cf5032B2a844602278b01199ED191A86c93ff",
-        "nonce": 0,
-        "value": "100000000000000000",
-        "gasLimit": 100000,
-        "gasPrice": "1000000000",
-        "chainId": 1001
       },
       {
         "id": 3,
@@ -367,7 +347,6 @@
     "expectedOldRoot": "0x090bcaf734c4f06c93954a827b45a6e8c67b8e0fd1e0a35a1c5982d6961828f9",
     "txs": [
       {
-        "encodeInvalidData": true,
         "id": 0,
         "from": "0x617b3a3528F9cDd6630fd3301B9c8911F7Bf063D",
         "to": "0x4d5Cf5032B2a844602278b01199ED191A86c93ff",
@@ -376,12 +355,12 @@
         "gasLimit": 100000,
         "gasPrice": "1000000000",
         "chainId": 1001,
-        "rawTx": "0xf86d80843b9aca00830186a0944d5cf5032b2a844602278b01199ed191a86c93ff88016345785d8a0000808207f5a01186622d03b6b8da7cf111d1ccba5bb185c56deae6a322cebc6dda0556f3cb979f910c26408b64b51c5da36ba2f38ef55ba1cee719d5a6c01225968799907432",
+        "rawTx": "0xf86e80843b9aca00830186a0944d5cf5032b2a844602278b01199ed191a86c93ff88016345785d8a0000808207f5a01186622d03b6b8da7cf111d1ccba5bb185c56deae6a322cebc6dda0556f3cb97a019683074bcc81dba07fad2ac4015cf2eba4807c1aa1a8d291e77317a45fc2023",
         "overwrite": {
-          "s": "0xa9683074bcc81dba07fad2ac4015cf2eba4807c1aa1a8d291e77317a45fc2023"
+          "s": "0x19683074bcc81dba07fad2ac4015cf2eba4807c1aa1a8d291e77317a45fc2023"
         },
-        "customRawTx": "0xee80843b9aca00830186a0944d5cf5032b2a844602278b01199ed191a86c93ff88016345785d8a0000808203e980801186622d03b6b8da7cf111d1ccba5bb185c56deae6a322cebc6dda0556f3cb9700910c26408b64b51c5da36ba2f38ef55ba1cee719d5a6c012259687999074321b",
-        "reason": "TX INVALID: Failed signature"
+        "customRawTx": "0xee80843b9aca00830186a0944d5cf5032b2a844602278b01199ed191a86c93ff88016345785d8a0000808203e980801186622d03b6b8da7cf111d1ccba5bb185c56deae6a322cebc6dda0556f3cb9719683074bcc81dba07fad2ac4015cf2eba4807c1aa1a8d291e77317a45fc20231b",
+        "reason": "TX INVALID: Not enough funds to pay total transaction cost"
       }
     ],
     "expectedNewRoot": "0x090bcaf734c4f06c93954a827b45a6e8c67b8e0fd1e0a35a1c5982d6961828f9",
@@ -395,11 +374,11 @@
         "nonce": "0"
       }
     },
-    "batchL2Data": "0xee80843b9aca00830186a0944d5cf5032b2a844602278b01199ed191a86c93ff88016345785d8a0000808203e980801186622d03b6b8da7cf111d1ccba5bb185c56deae6a322cebc6dda0556f3cb9700910c26408b64b51c5da36ba2f38ef55ba1cee719d5a6c0122596879990",
+    "batchL2Data": "0xee80843b9aca00830186a0944d5cf5032b2a844602278b01199ed191a86c93ff88016345785d8a0000808203e980801186622d03b6b8da7cf111d1ccba5bb185c56deae6a322cebc6dda0556f3cb9719683074bcc81dba07fad2ac4015cf2eba4807c1aa1a8d291e77317a45fc20231b",
     "globalExitRoot": "0x090bcaf734c4f06c93954a827b45a6e8c67b8e0fd1e0a35a1c5982d6961828f9",
     "newLocalExitRoot": "0x17c04c3760510b48c6012742c540a81aba4bca2f78b9d14bfd2f123e2e53ea3e",
-    "inputHash": "0x2b683d3733de4400656d108e35d9a8ac2ccd11df439c48019eb1b5f7278bcf07",
-    "batchHashData": "0x666d0ee24aac702077a753387f90f4d9c3004a6acec7cad44097fa36987aed64",
+    "inputHash": "0x0c69be3e142d0008c051a7f584e44a8049bf4db323881751a81782889d0ae80a",
+    "batchHashData": "0xb9c0cb7695d04dd27b85bb2a78845f14adeab04d91c3d959b92cde3127b4959f",
     "localExitRoot": "0x17c04c3760510b48c6012742c540a81aba4bca2f78b9d14bfd2f123e2e53ea3e",
     "timestamp": 1944498031
   },
